anv: Use the correct image layout for self-dep input attachments
Annoyingly, there's no good way to track this information except to pull
the input attachment <-> binding map from the shader source.  We could
pull a generic input attachment index from the set instead of poking
into the render pass but then we'd have to poke at the render pass at
draw time and that's not much better of a situation.