HACK: anv: Flush the tile cache when re-enabling CCS
The following sequence causes problems for unknown reasons:

 - Render with CCS enabled
 - Full CCS resolve
 - RT and Tile cache flush
 - Render without CCS, sampling from render target without CCS
 - RT flush and sampler invalidate (no tile cache flush)
 - Sample from the image with CCS

The data read by the last operation is incorrect.  In particular, it
reads the data written by the first draw, not the second, as if it's
skipping the tile cache and reading straight from main memory.  Either
of the following seem to fix the issue:

 - Disable CCS on the final sampling operation

OR

 - Flush the tile cache between rendering and sampling along with the RT
   flush described above

It's unclear whether it's ther ender hardware or the sampler that's
skipping the tile cache when CCS is disabled but one of them is.