drisw: Don't expose __DRI_DRI2
dri2_load_driver is only called from the actual X11 DRI2 protocol path,
which is the only place this extension could bind. xserver typically
does not expose DRI2 when the GL under it is swrast, and even if it did,
we do not (for example) use DRI2 to allocate llvmpipe's winsys buffers,
so there's no benefit to handling swkms specially here.