dri: Use the common dri2 format mapping instead of driGLFormatToImageFormat().
This means that we don't try to map L8 to DRI_IMAGE_FORMAT_R8 (and
L8A8->R8G8) any more, but we shouldn't be doing that because a map on the
other side would turn that into R8G instead of L8.  Still passes
dEQP-EGL.functional.image.api.create_image.*, which is what I think the
L8->R8 mapping was introducd for.

Reviewed-by: Adam Jackson <ajax@redhat.com>
Reviewed-by: Marek Olšák <marek.olsak@amd.com>