intel/fs: Skip discarding the interference graph on most hardware
Only Gfx7-8 require reserving registers for constructing spilling
messages.  Gfx4-6 have actual MRFs for this purpose, and Gfx9+ has
split sends and constructs message headers in new temporary registers
with a very short lifespan which are simply added to the existing
interference graph as new nodes and allocated via the normal mechanism.

This means that when we need to spill for the first time, we can avoid
discarding and recomputing the entire interference graph.  We also avoid
a doomed ra_allocate() pass needed simply to recreate spill candidate
information on the new graph that has additional interference.  Because
the graph remains valid and none of the existing nodes had any changes
to their interference, the existing spill candidates remain valid.

This will slightly help improve compile time when needing to spill.