freedreno: Add default highest_bank_bit values
This value is programmed by the kernel and not determined by the
hardware, but provide a default value that should match what drm/msm
programs for older kernels that can't report it. kgsl has always
supported returning the highest_bank_bit, so we don't have to worry
about it.

Note that this changed recently for a690, so it may be broken there
without a bleeding-edge kernel, but if you're not carrying that change
then things are generally already pretty broken.