panvk: Unconditionally invalidate descriptor tables in CmdBindDescriptorSets()
We can't invalidate some of the tables only if we have descriptors
of this type in the new set, because following sets might be impacted
by those changes too.

We could be smart and invalidate only if the old/new sets have a
different layout, but let's keep things simple for now, and invalidate
everything to force descriptor table emission.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>