panvk: Move translate_sampler_compare_func() to panvk_vX_sampler.c
Won't be used outside the sampler logic, so we'd rather have this
function moved to the panvk_vX_sampler.c.

We replace the translate_compare_func() call by a simple cast, the
STATIC_ASSERT() in there should guarantee we always have a valid
mali_func value. This allows us to get rid of the panck_cs.h
inclusion.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>