venus: Use common physical device properties
This lets us delegate the GPDP2 entrypoint to common code.
This also lets us delete struct vn_physical_device_properties,
as it is redundant with struct vk_properties present in the runtime.

Move the properties present in vn_physical_device_properties to the
local_devices struct used to query the host device properties, so we can
still get and fill those properties. Remove the extension checking
logic, and fill all the properties regardless. (the structs are
initalized to zero anyway, so this should not be a problem.)

Replace accesses to struct vn_physical_device with accesses to
struct vk_properties filled in at device initialization time.