ac/sqtt, radv: extract helpers from radv
This commit adds the ac_prepare_sqtt_start / ac_prepare_sqtt_stop
helpers and use them from radeonsi.

The goal is to reduce code duplication as much as possible. Note
that since writting to cs is handled slightly differently by
the drivers, it's not really possible to pass the target buffer
directly.

Reviewed-by: Marek Olšák <marek.olsak@amd.com>