anv: Always expose a compute only queue when supported by HW and KMD
Even if we can't implement compute queue with a compute engine because
of TRTT, we can implement it render engine.

This is necessary to have the same number of queues when application
uses Sparse and when it don't, the next patches allow to implement
this queue with compute engine when sparse it not in use.

This approach could also be used to enable a transfer only queue but
tests are failling when a transfer only queue is implemented with
render engine, so leaving it as TODO.

Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>