anv/trtt: also program the BLT and COMP_CTX0 TRTT registers
This is enough to make TR-TT work on the compute and blitter engines
on my LNL machine.

Notice that on the BSpec pages for the TR-TT registers, the "Register
Instance" section only lists the GFX registers as non-priviledged.
However, the "User Mode Privileged Commands" lists the other instances
of the TR-TT Regsiters as non-privildeged, which matches what we see:
there's no need to put these addresses in the FORCE_TO_NONPRIV
registers.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>