RFC: anv/trtt: enable TR-TT for compute and blitter on gen 12 too
   DISCLAIMER: this is marked as an RFC since apparently some sparse
               tests RANDOMLY fail when this is enabled!

The BSpec page for Tiled Resources Translation Tables for gen 12
(44990) says:
  "Usage model for TR translations are restricted to GFX Render Engine
   (& POSH pipeline)."

Despite that, treating it the same as we treat info.ver20 makes it
work, so let's try enabling it.

Notice that on the BSpec pages for the TR-TT registers, the "Register
Instance" section only lists the GFX registers as non-priviledged.
However, the "User Mode Privileged Commands" lists the other instances
of the TR-TT Regsiters as non-privildeged, which matches what we see:
there's no need to put these addresses in the FORCE_TO_NONPRIV
registers.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>