tu: use either the 16-bit or 32-bit descriptor
Until now, if the 16-bit storage functionality is supported by the
hardware, two separate descriptors were set up, with isam loads and stores
piping through the descriptor of the corresponding size and other storage
access using the 16-bit descriptor.

These changes keep separate descriptors on a650, but leverage post-a650
functionality that enables isam to use 16-bit descriptors as well, removing
the need for the separate 32-bit descriptor.

A new device property variable is provided, enabled on 4th-gen a6xx devices
and onwards. Storage buffer descriptors are then set up according to 16-bit
storage support and the indicated capability of using those descriptors for
32-bit isam loads as well.

Dynamic offset application in tu_CmdBindDescriptorSets is modified to
determine the offset shift value based on the descriptor's format and not
on the descriptor's position in the layout binding.

Signed-off-by: Zan Dobersek <zdobersek@igalia.com>