anv/xe: try harder when the vm_bind ioctl returns ENOMEM
From all the many possible errors returned by the vm_bind ioctl,
ENOMEM is the one that can actually happen in the wild, when the
system is under memory pressure. Thomas Hellström also pointed us
that, due to its asynchronous nature, the vm_bind ioctl itself has to
pin some memory, so if the number of bind operations passed is too
big, the probability that it may run out of memory increases.

So in this patch, when we detect the system is under memory pressure,
we throw away our performance expectations and try to go slowly and
steady. First we wait everything we're supposed to wait, and then we
synchronously bind one piece at a time, hoping that this won't cause
the Kernel to try to reserve too much memory. All this while also
hoping that whatever thing that may be eating all the memory goes away
in the meantime. If even this fails, we give up and hope the upper
layer will be able to figure out what to do.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>