etnaviv: isa: Extend disasm test
With libetnaviv_parser we are able to parse the resulting string
representation into an etna_inst and assemble that to binary.

As we are not able to parse and/or assemble we need to mark some test
cases with special flags.

This allows us to test: bin -> disasm -> parsing -> assemble

Signed-off-by: Christian Gmeiner <cgmeiner@igalia.com>