WIP: tool/mda: Add tool to inspect mda.* files
The mda.* files are regular tar files, so can be inspected by tar,
and also used direct by file managers and editors.  However a few
common tasks are worth having already set up in the repository.

This tool adds some of them: showing the last state of an object (shader
representation), showing the sequence of states, showing a log of the
changes between the states, showing the diff between two states of a
represenation.  See the "help message" inside the commit for details.