panvk: Move panvk_pipeline_layout state to panvk_set_collection_layout
This move all state that previously was previous stored in
panvk_pipeline_layout to a new struct panvk_set_collection_layout.

This also add various properties that will be used for command buffers.

This is required to remove dependency on the pipeline layout in the
shader compilation path in preparation for vk_shader migration.

Signed-off-by: Mary Guillemard <mary.guillemard@collabora.com>